requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
faiss==1.9.0
faiss-cpu==1.10.0
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-anthropic==0.3.10
langchain-community==0.3.21
langchain-core==0.3.51
langchain-deepseek==0.1.3
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langsmith==0.3.23
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
openai==1.70.0
orjson==3.10.16
# packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pdf2image==1.17.0
pillow==11.1.0
pluggy==1.5.0
propcache==0.3.1
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyJWT==2.10.1
pymongo==4.11.3
pytesseract==0.3.13
pytest==8.3.5
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
setuptools==75.8.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
sympy==1.13.1
tenacity==9.1.2
textblob==0.19.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
# torch==2.6.0+cu118
# torchaudio==2.6.0+cu118
torchvision==0.21.0
tqdm==4.67.1
transformers==4.50.3
twilio==9.5.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
wheel==0.45.1
yarl==1.18.3
zstandard==0.23.0
